# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DialogueHistoryResponse": {
        "description": "某個靈魂的對話，`turns` **由舊到新**（畫面由上往下讀）。\n\n分頁往「更舊」的方向走：把 `next_before` 原封不動放進下一次請求的 `before`。",
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/DialogueTurnItem"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "turns",
          "has_more"
        ],
        "title": "DialogueHistoryResponse",
        "type": "object"
      },
...
      "DialogueThreadItem": {
        "description": "聊天紀錄第一層：跟這個靈魂的對話摘要。",
        "properties": {
          "last_at": {
            "format": "date-time",
            "title": "Last At",
            "type": "string"
          },
          "last_message": {
            "title": "Last Message",
            "type": "string"
          },
          "last_role": {
            "enum": [
              "player",
              "spirit"
            ],
            "title": "Last Role",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "spirit_id": {
            "title": "Spirit Id",
            "type": "string"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          }
        },
        "required": [
          "spirit_id",
          "name",
          "last_message",
          "last_role",
          "last_at",
          "turn_count"
        ],
        "title": "DialogueThreadItem",
        "type": "object"
      },
      "DialogueThreadsResponse": {
        "description": "一句都沒聊過時 `threads` 是空陣列，不是 404——冷啟動是正常狀態。",
        "properties": {
          "threads": {
            "items": {
              "$ref": "#/components/schemas/DialogueThreadItem"
            },
            "title": "Threads",
            "type": "array"
          }
        },
        "required": [
          "threads"
        ],
        "title": "DialogueThreadsResponse",
        "type": "object"
      },
      "DialogueTurnItem": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "role": {
            "enum": [
              "player",
              "spirit"
            ],
            "title": "Role",
            "type": "string"
          },
          "turn_id": {
            "title": "Turn Id",
            "type": "integer"
          }
        },
        "required": [
          "turn_id",
          "role",
          "content",
          "created_at"
        ],
        "title": "DialogueTurnItem",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "所有錯誤回應的統一形狀（SDD v2.1 §11.2.1）。\n\n**這不是新的錯誤格式，是對既有格式的描述。** 形狀刻意對齊 FastAPI\n`HTTPException` 本來就會產出的 `{\"detail\": \"...\"}`，一個位元組都沒改。\n\n存在的理由是讓契約說得出錯誤長什麼樣。在這之前，錯誤回應完全不在 OpenAPI\n文件裡——SDD v2.1 §10.3 明訂「錯誤碼全數不變」代表錯誤碼是契約的一部分，\n但機器可讀的那份契約看不到它們，客戶端只能回頭讀規格散文，這正好違背\n「單一真相來源」的初衷。\n\n具名之後，客戶端 codegen 會產出**單一**錯誤 DTO，錯誤處理可以寫成一份共用\n程式碼，而不是每支端點各寫一次。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResonanceProgressResponse": {
        "description": "`GET /api/v1/resonance/{spiritId}` 的回應（SDD v1 §8.9）。\n\n`stage` 與 `next_threshold` 一律由 `resonance_value` 重算。\n已滿階時 `next_threshold` 為 null。",
        "properties": {
          "next_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Threshold"
          },
          "resonance_value": {
            "title": "Resonance Value",
            "type": "integer"
          },
          "spirit_id": {
            "title": "Spirit Id",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "integer"
          }
        },
        "required": [
          "spirit_id",
          "resonance_value",
          "stage"
        ],
        "title": "ResonanceProgressResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "城市靈魂 AR — Backend",
    "version": "0.1.0-sprint1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dialogue-history/spirits": {
      "get": {
        "description": "聊天紀錄第一層：我跟哪些靈魂聊過。\n\n🔒 **只回自己的。** 過濾條件是 session token 解出來的 `player_id`，沒有任何\n路徑或查詢參數能指定別人——這是設計，不是驗證不足。",
        "operationId": "list_dialogue_threads_api_v1_dialogue_history_spirits_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DialogueThreadsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session token 無效或未提供"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Dialogue Threads",
        "tags": [
          "body"
        ]
      }
    },
...
    "/api/v1/resonance/{spirit_id}": {
      "get": {
        "description": "S8．共鳴進度查詢（#35）。\n\n「還沒開始」回 0 而不是 404——前端要能直接畫一條 0/10 的進度條。\n地標不存在或已下架才是 404。",
        "operationId": "get_resonance_api_v1_resonance__spirit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "spirit_id",
            "required": true,
            "schema": {
              "title": "Spirit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResonanceProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session token 無效或未提供"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "靈魂不存在，或已下架（is_active=false）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Resonance",
        "tags": [
          "body"
        ]
      }
    },
...
    "/api/v1/spirits/{place_id}/dialogue-history": {
      "get": {
        "description": "聊天紀錄第二層：跟某個靈魂的對話內文，由舊到新。\n\n## 為什麼靈魂不存在要回 404，而不是空陣列\n\n「這個靈魂我沒聊過」與「根本沒有這個靈魂」是兩種不同的情況，客戶端要能分開：\n前者顯示「還沒有聊過」，後者代表客戶端拿著一個過期的 id 在問——那是 bug，\n不該被偽裝成一個平靜的空畫面。對齊 `GET /resonance/{spirit_id}` 的做法。\n\n下架的靈魂也回 404（跟其他所有靈魂查詢端點一致）。**但已經存在的對話紀錄\n本身不會被藏起來**，它們仍然出現在第一層的清單裡——見\n`dialogue_log.threads()` 的說明。",
        "operationId": "get_dialogue_history_api_v1_spirits__place_id__dialogue_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "place_id",
            "required": true,
            "schema": {
              "title": "Place Id",
              "type": "string"
            }
          },
          {
            "description": "一次回幾句。",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "一次回幾句。",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "只回 turn_id 比這個小的（更舊的）。第一頁不要帶。",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "只回 turn_id 比這個小的（更舊的）。第一頁不要帶。",
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DialogueHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session token 無效或未提供"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "靈魂不存在，或已下架（is_active=false）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Dialogue History",
        "tags": [
          "body"
        ]
      }
    },
...
  }
}